update room create

+++ back-end/app-back-end/src/controllers/co-working.controller.ts
@@ -264,12 +264,14 @@ export class CoWorkingController {
         const coWorking = await this.coWorkingRepository.findById(id, {
             include: [{relation: 'rooms'}],
         });
-        for (let r of coWorking.rooms) {
-            const room = await this.roomRepository.deleteRoom(r.id);
+        console.log(coWorking);
+        if (coWorking.rooms) {
+            for (let r of coWorking.rooms) {
+                const room = await this.roomRepository.deleteRoom(r.id);
+            }
         }
         delete coWorking.rooms;
         deleteFiles(coWorking.photo);
-        console.log(coWorking);
         await this.coWorkingRepository.delete(coWorking);
         // console.log(coWorking);
     }
